[ALPS05712676] APE_DEC: fix coverity
[ALPS05713413] APE_DEC: fix coverity issue

Fix coverity issue for new ape decoder.

[Verification details]
Verified project:
k6893v1_64_swrgo_userdebug
Did you build pass? (Yes/No)
Yes
Did you boot up? (Yes/No)
Yes

MTK-Commit-Id: bb50dd7b7479a5853bdd48312a85a51553d1ee66

CR-Id: ALPS05712676
Feature: APE Playback

+++ src/ape_decoder_swip.c
@@ -513,7 +513,7 @@ static inline int ape_decode_value_3900(APEContext *ctx, APERice *rice)
     update_rice(rice, x);
 
     /* Convert to signed */
-    return ((x >> 1) ^ ((x & 1) - 1)) + 1;
+    return (int)(((x >> 1) ^ ((x & 1) - 1)) + 1);
 }
 
 static inline int ape_decode_value_3990(APEContext *ctx, APERice *rice)
@@ -548,7 +548,7 @@ static inline int ape_decode_value_3990(APEContext *ctx, APERice *rice)
     x = base + overflow * pivot;
     update_rice(rice, x);
     /* Convert to signed */
-    return ((x >> 1) ^ ((x & 1) - 1)) + 1;
+    return (int)(((x >> 1) ^ ((x & 1) - 1)) + 1);
 }
 
 static void decode_array_0000(APEContext *ctx, GetBitContext *gb,
